Update alphaRouter utility function, modularize code
